stream/requirements.txt
flask 
kafka-python 
gunicorn